Move files and conftest
