+++ py-polars/tests/unit/io/files/gzipped.csv
