+++ py-polars/tests/unit/io/files/small.avro
